requirements.txt
Django==6.0.4
djangorestframework
djangorestframework-simplejwt
django-filter
python-dotenv==1.1.0
mysqlclient==2.2.8
